// src/types/types.ts
export interface UserRegistration {
    email: string;
    firstname: string;
    lastname: string;
    nickname: string;
    password: string;
    password2: string;
    captcha: string | null;
}